NSMPnames.hgu133plus2(panp)	R Documentation

Negative Strand Matching Probeset (NSMP) names for HG-U133 Plus 2.0 chip type

Description

This list of NSMP probeset names is used by pa.calls() in calculating the NSMP expression distribution. It is specific to the chip type; the correct version is automatically loaded for the detected chip type, if supported. The user does not load or use this data directly.

Usage

data(NSMPnames.hgu133plus2)

Format

A vector containing probeset names

Source

October 2004 AffyMetrix chip annotation files, filtered, with outliers removed

[Package panp version 1.0.0 Index]