gcrma.exprSet(panp)	R Documentation

ExprSet resulting from gcrma processing of 3 HG-U133A chips

Description

An exprSet object resulting from gcrma processing of three HG-U133A chips from the 28-chip affyCOMP II dataset, derived from the AffyMetrix Latin Squares dataset. The gcrma.exprSet can be used as example input for panp's pa.calls() function, and it is used for this purpose in the panp vignette.

Usage

data(gcrma.exprSet)

Format

An exprSet object comprising three gcrma-processed HG-U133A chips

Source

The 3 HG-U133A chips are the first three of affyCOMP II Latin Squares set

[Package panp version 1.0.0 Index]