test133a(panp)	R Documentation

Test HG-U133A exprSet for pa.calls

Description

An example exprSet object from a small set of HG-U133A chips, preprocessed with RMA. The pa.calls() function can be executed on this exprSet, to exercise the function and generate P/A calls and related P-values. See ?help(pa.calls) for details. This exprSet is the one used in the help(pa.calls) example.

Usage

data(test133a)

Format

A 3-chip HG-U133A exprSet object (see the affy package for details on this object type) containing 3 samples, 22300 genes

Source

A subset of the Latin Squares samples used in the affyCOMP II package

[Package panp version 1.0.0 Index]